объединение модулей comments-gen. и post-gen в один data.js
