+++ js/main.js
@@ -1,4 +1,4 @@
-import {simularPhotoList} from './post-generation.js';
+import {simularPhotoList} from './data.js';
 
 // eslint-disable-next-line no-console
 console.table(simularPhotoList());
